refactor: Preview Data 통합
- 각 Component에 맞게 분리해둔 Preview 함수를 하나의 공간에 묶음
