+++ app/src/main/java/com/woowa/accountbook/ui/calendar/component/CalendarTotText.kt
@@ -7,9 +7,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.woowa.accountbook.ui.theme.AccountbookTheme
 import com.woowa.accountbook.ui.theme.Purple200
 import com.woowa.accountbook.ui.theme.Purple700
+import com.woowa.accountbook.ui.theme.Teal200
 
 @Composable
 fun CalendarTotText(
@@ -39,5 +42,13 @@ fun CalendarTotText(
                 .height(1.dp),
             color = Purple200,
         )
+    }
+}
+
+@Preview
+@Composable
+fun CalendarTotTextPreview() {
+    AccountbookTheme {
+        CalendarTotText("수입", "123122321", color = Teal200)
     }
 }
