+++ app/src/main/java/com/woowa/accountbook/ui/common/component/ContentWithTitleItem.kt
@@ -11,8 +11,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
+import com.woowa.accountbook.ui.theme.AccountbookTheme
 
 @Composable
 fun ContentWithTitleItem(
@@ -36,5 +38,13 @@ fun ContentWithTitleItem(
             content()
         }
         SubDivider()
+    }
+}
+
+@Preview
+@Composable
+fun ContentWithTitleItemPreview() {
+    AccountbookTheme {
+        ContentWithTitleItem(title = "Test")
     }
 }
