+++ app/src/main/java/com/woowa/accountbook/ui/common/component/MainAppBar.kt
@@ -13,8 +13,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.woowa.accountbook.R
+import com.woowa.accountbook.ui.theme.AccountbookTheme
 
 @Composable
 fun MainAppBar(
@@ -68,5 +70,13 @@ fun MainAppBar(
             color = contentColor,
             thickness = 1.dp,
         )
+    }
+}
+
+@Preview
+@Composable
+fun MainAppBarPreview() {
+    AccountbookTheme {
+        MainAppBar(title = "Preview!")
     }
 }
