+++ app/src/main/java/com/woowa/accountbook/ui/common/component/TextFieldWithHint.kt
@@ -5,13 +5,18 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.TextLayoutResult
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
+import com.woowa.accountbook.ui.theme.AccountbookTheme
 
 @Composable
 fun TextFieldWithHint(
@@ -52,5 +57,24 @@ fun TextFieldWithHint(
             visualTransformation = visualTransformation
         )
         if (value.isEmpty()) hint()
+    }
+}
+
+@Preview
+@Composable
+fun TextFieldWithHintPreview() {
+    AccountbookTheme {
+        TextFieldWithHint(
+            "1234",
+            onValueChange = { str -> },
+            singleLine = true,
+            modifier = Modifier.fillMaxWidth(),
+            textStyle = TextStyle(
+                color = MaterialTheme.colors.primary,
+                fontWeight = FontWeight.Bold
+            ),
+            keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.NumberPassword),
+            keyboardActions = KeyboardActions(),
+        )
     }
 }
