+++ app/src/main/java/com/woowa/accountbook/ui/history/component/HistoryExpenditureFilterButton.kt
@@ -12,8 +12,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.woowa.accountbook.R
+import com.woowa.accountbook.ui.theme.AccountbookTheme
 import com.woowa.accountbook.ui.theme.Purple200
 import com.woowa.accountbook.utils.StringUtil
 
@@ -51,5 +53,13 @@ fun HistoryExpenditureFilterButton(
         Text(text = "  지출  ")
         if (itemVisible)
             Text(text = StringUtil.getMoneyFormatString(totPrice))
+    }
+}
+
+@Preview
+@Composable
+fun HistoryExpenditureFilterButtonPreview() {
+    AccountbookTheme {
+        HistoryExpenditureFilterButton(isChecked = true, isActive = true)
     }
 }
