+++ app/src/main/java/com/woowa/accountbook/ui/history/component/HistoryMainFilterButton.kt
@@ -9,7 +9,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.woowa.accountbook.ui.theme.AccountbookTheme
 
 @Composable
 fun HistoryMainFilterButton(
@@ -47,5 +49,13 @@ fun HistoryMainFilterButton(
             itemVisible = itemVisible,
             totPrice = totExpenditure
         )
+    }
+}
+
+@Preview
+@Composable
+fun HistoryMainFilterButtonPreview() {
+    AccountbookTheme {
+        HistoryMainFilterButton()
     }
 }
