+++ app/src/main/java/com/woowa/accountbook/ui/history/manage/ManageHistoryFragment.kt
@@ -189,7 +189,7 @@ class ManageHistoryFragment : Fragment() {
                     Box(
                         modifier = Modifier
                             .align(Alignment.BottomCenter)
-                            .padding(vertical = 40.dp)
+                            .padding(vertical = 40.dp, horizontal = 16.dp)
                     ) {
                         CommonButton(
                             text = if (editFlag) "수정하기" else "등록하기",
