--- app/src/main/java/com/woowa/accountbook/ui/history/manage/NewHistoryPreviews.kt
@@ -1,10 +1,0 @@
-package com.woowa.accountbook.ui.history.manage
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-
-@Composable
-@Preview
-fun CountrySelectionPreview() {
-    /*DropDownComponent()*/
-}
