--- app/src/main/java/com/woowa/accountbook/ui/setting/SettingPreviews.kt
@@ -1,13 +1,0 @@
-package com.woowa.accountbook.ui.setting
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.woowa.accountbook.ui.theme.AccountbookTheme
-
-@Composable
-@Preview
-fun SettingMainItemPreviews() {
-    AccountbookTheme {
-
-    }
-}
