+++ app/src/main/java/com/woowa/accountbook/ui/setting/manage/component/ColorPaletteComponent.kt
@@ -3,17 +3,16 @@ package com.woowa.accountbook.ui.setting.manage.component
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.defaultMinSize
-import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.GridCells
 import androidx.compose.foundation.lazy.LazyVerticalGrid
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.woowa.accountbook.ui.theme.*
 
 @ExperimentalFoundationApi
 @Composable
@@ -43,4 +42,47 @@ fun ColorPaletteComponent(
             )
         }
     }
+}
+
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+@Preview
+fun ColorPaletteComponentPreviews() {
+    ColorPaletteComponent(
+        listOf(
+            Blue1,
+            Blue2,
+            Blue3,
+            Blue4,
+            Blue5,
+            Green1,
+            Green2,
+            Green3,
+            Green4,
+            Green5,
+            Purple1,
+            Purple2,
+            Purple3,
+            Purple4,
+            Purple5,
+            Pink1,
+            Pink2,
+            Pink3,
+            Pink4,
+            Pink5,
+            Olive1,
+            Olive2,
+            Olive3,
+            Olive4,
+            Olive5,
+            Yellow1,
+            Yellow2,
+            Yellow3,
+            Yellow4,
+            Yellow5
+        ),
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp)
+    )
 }
