+++ app/src/main/java/com/woowa/accountbook/ui/statistic/component/GraphComponent.kt
@@ -7,10 +7,12 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.woowa.accountbook.domain.model.CategoryStatisticDto
 import com.woowa.accountbook.domain.model.toSliceOfPieChartData
+import com.woowa.accountbook.ui.theme.AccountbookTheme
 import me.bytebeats.views.charts.pie.PieChart
 import me.bytebeats.views.charts.pie.PieChartData
 import me.bytebeats.views.charts.pie.render.SimpleSliceDrawer
@@ -35,6 +37,20 @@ fun GraphComponent(
                 .align(Alignment.Center),
             animation = simpleChartAnimation(),
             sliceDrawer = SimpleSliceDrawer(thickness)
+        )
+    }
+}
+
+@Composable
+@Preview
+fun GraphComponentPreview() {
+    AccountbookTheme {
+        GraphComponent(
+            listOf(
+                /*CategoryStatisticDto(Category(), 12345, 0.3f),
+                CategoryStatisticDto(Category(), 12345, 0.3f),
+                CategoryStatisticDto(Category(), 12345, 0.4f)*/
+            )
         )
     }
 }
